Revert "Feat/store/#36 store delete api"
