+++ store-service/build.gradle
@@ -25,7 +25,6 @@ repositories {
 
 ext {
     set('springCloudVersion', "2024.0.1")
-    set('querydslVersion', "5.0.0")
 }
 
 dependencies {
@@ -39,13 +38,7 @@ dependencies {
     implementation 'org.springframework.cloud:spring-cloud-starter-netflix-eureka-client'
 
     // common
-    implementation 'io.github.boeingmerryho:common-library:1.0.4'
-
-    // querydsl
-    implementation "com.querydsl:querydsl-jpa:${querydslVersion}:jakarta"
-    annotationProcessor "com.querydsl:querydsl-apt:${querydslVersion}:jakarta"
-    annotationProcessor "jakarta.annotation:jakarta.annotation-api"
-    annotationProcessor "jakarta.persistence:jakarta.persistence-api"
+    implementation 'io.github.boeingmerryho:common-library:1.0.3'
 
     // map struct
     implementation "org.mapstruct:mapstruct:1.5.5.Final"
@@ -59,12 +52,6 @@ dependencies {
     // Database
     runtimeOnly 'com.mysql:mysql-connector-j'
     implementation 'org.springframework.boot:spring-boot-starter-jdbc'
-
-    // Redis
-    implementation 'org.springframework.boot:spring-boot-starter-data-redis'
-
-    // Kafka
-    implementation 'org.springframework.kafka:spring-kafka'
 
     // test
     testImplementation 'org.springframework.boot:spring-boot-starter-test'
