+++ store-service/src/main/java/com/boeingmerryho/business/storeservice/StoreServiceApplication.java
@@ -3,11 +3,9 @@ package com.boeingmerryho.business.storeservice;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
 @EnableDiscoveryClient
-@EnableScheduling
 public class StoreServiceApplication {
 
 	public static void main(String[] args) {
