--- store-service/src/main/java/com/boeingmerryho/business/storeservice/application/dto/query/StoreSearchCondition.java
@@ -1,9 +1,0 @@
-package com.boeingmerryho.business.storeservice.application.dto.query;
-
-public record StoreSearchCondition(
-	Long stadiumId,
-	String name,
-	Boolean isClosed,
-	Boolean isDeleted
-) {
-}
