--- store-service/src/main/java/com/boeingmerryho/business/storeservice/application/dto/request/StoreDetailRequestServiceDto.java
@@ -1,6 +1,0 @@
-package com.boeingmerryho.business.storeservice.application.dto.request;
-
-public record StoreDetailRequestServiceDto(
-	Long id
-) {
-}
