--- store-service/src/main/java/com/boeingmerryho/business/storeservice/application/dto/request/StoreSearchRequestServiceDto.java
@@ -1,11 +1,0 @@
-package com.boeingmerryho.business.storeservice.application.dto.request;
-
-import org.springframework.data.domain.Pageable;
-
-public record StoreSearchRequestServiceDto(
-	Pageable customPageable,
-	Long stadiumId,
-	String name,
-	Boolean isClosed
-) {
-}
