--- store-service/src/main/java/com/boeingmerryho/business/storeservice/application/dto/request/StoreUpdateRequestServiceDto.java
@@ -1,10 +1,0 @@
-package com.boeingmerryho.business.storeservice.application.dto.request;
-
-import java.time.LocalDateTime;
-
-public record StoreUpdateRequestServiceDto(
-	String name,
-	LocalDateTime openAt,
-	LocalDateTime closedAt
-) {
-}
