--- store-service/src/main/java/com/boeingmerryho/business/storeservice/application/dto/response/StoreDetailAdminResponseServiceDto.java
@@ -1,15 +1,0 @@
-package com.boeingmerryho.business.storeservice.application.dto.response;
-
-import java.time.LocalDateTime;
-
-public record StoreDetailAdminResponseServiceDto(
-	Long id,
-	Long stadiumId,
-	String name,
-	LocalDateTime openAt,
-	LocalDateTime closedAt,
-	Boolean isClosed,
-	Boolean isDeleted,
-	Boolean isQueueAvailable
-) {
-}
