--- store-service/src/main/java/com/boeingmerryho/business/storeservice/application/dto/response/StoreDetailResponseServiceDto.java
@@ -1,14 +1,0 @@
-package com.boeingmerryho.business.storeservice.application.dto.response;
-
-import java.time.LocalDateTime;
-
-public record StoreDetailResponseServiceDto(
-	Long id,
-	Long stadiumId,
-	String name,
-	LocalDateTime openAt,
-	LocalDateTime closedAt,
-	Boolean isClosed,
-	Boolean isQueueAvailable
-) {
-}
