--- store-service/src/main/java/com/boeingmerryho/business/storeservice/application/dto/response/StoreSearchResponseServiceDto.java
@@ -1,13 +1,0 @@
-package com.boeingmerryho.business.storeservice.application.dto.response;
-
-import java.time.LocalDateTime;
-
-public record StoreSearchResponseServiceDto(
-	Long id,
-	Long stadiumId,
-	String name,
-	LocalDateTime openAt,
-	LocalDateTime closedAt,
-	Boolean isClosed
-) {
-}
