+++ store-service/src/main/java/com/boeingmerryho/business/storeservice/application/service/StoreAdminService.java
@@ -1,20 +1,13 @@
 package com.boeingmerryho.business.storeservice.application.service;
 
-import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
 
 import com.boeingmerryho.business.storeservice.application.dto.mapper.StoreApplicationMapper;
 import com.boeingmerryho.business.storeservice.application.dto.request.StoreCreateRequestServiceDto;
-import com.boeingmerryho.business.storeservice.application.dto.request.StoreSearchAdminRequestServiceDto;
-import com.boeingmerryho.business.storeservice.application.dto.request.StoreUpdateRequestServiceDto;
 import com.boeingmerryho.business.storeservice.application.dto.response.StoreCreateResponseServiceDto;
-import com.boeingmerryho.business.storeservice.application.dto.response.StoreDetailAdminResponseServiceDto;
-import com.boeingmerryho.business.storeservice.application.dto.response.StoreSearchAdminResponseServiceDto;
-import com.boeingmerryho.business.storeservice.application.dto.response.StoreUpdateResponseServiceDto;
 import com.boeingmerryho.business.storeservice.domain.entity.Store;
-import com.boeingmerryho.business.storeservice.infrastructure.helper.StoreAdminHelper;
-import com.boeingmerryho.business.storeservice.infrastructure.helper.StoreQueueAdminHelper;
-import com.boeingmerryho.business.storeservice.infrastructure.helper.StoreValidator;
+import com.boeingmerryho.business.storeservice.domain.service.StoreDomainService;
+import com.boeingmerryho.business.storeservice.domain.service.StoreValidator;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -23,62 +16,15 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class StoreAdminService {
 
-	private final StoreAdminHelper storeAdminHelper;
+	private final StoreDomainService storeDomainService;
 	private final StoreApplicationMapper mapper;
 	private final StoreValidator validator;
-	private final StoreQueueAdminHelper storeQueueAdminHelper;
 
 	@Transactional
 	public StoreCreateResponseServiceDto createStore(StoreCreateRequestServiceDto requestServiceDto) {
 		validator.validateNotDuplicated(requestServiceDto.stadiumId(), requestServiceDto.name());
-
-		Store saved = storeAdminHelper.save(requestServiceDto);
+		Store saved = storeDomainService.save(requestServiceDto);
 		return mapper.toStoreCreateResponseServiceDto(saved);
 	}
 
-	public StoreDetailAdminResponseServiceDto getStoreDetail(Long id) {
-		Store storeDetail = storeAdminHelper.getAnyStoreById(id);
-
-		boolean isQueueAvailable = storeQueueAdminHelper.isQueueAvailable(storeDetail.getId());
-
-		return mapper.toStoreDetailAdminResponseServiceDto(storeDetail, isQueueAvailable);
-	}
-
-	public Page<StoreSearchAdminResponseServiceDto> searchStore(
-		StoreSearchAdminRequestServiceDto requestServiceDto) {
-		Page<Store> stores = storeAdminHelper.search(requestServiceDto);
-		return stores.map(mapper::toStoreSearchAdminResponseServiceDto);
-	}
-
-	@Transactional
-	public StoreUpdateResponseServiceDto updateStore(Long id, StoreUpdateRequestServiceDto requestDto) {
-		validator.validateHasUpdatableFields(requestDto);
-		Store updated = storeAdminHelper.updateStoreInfo(id, requestDto);
-		return mapper.toStoreUpdateResponseServiceDto(updated);
-	}
-
-	@Transactional
-	public StoreUpdateResponseServiceDto openStore(Long id) {
-		Store opened = storeAdminHelper.updateStoreOpen(id);
-		return mapper.toStoreUpdateResponseServiceDto(opened);
-	}
-
-	@Transactional
-	public StoreUpdateResponseServiceDto closeStore(Long id) {
-		Store closed = storeAdminHelper.updateStoreClose(id);
-		return mapper.toStoreUpdateResponseServiceDto(closed);
-	}
-
-	public void enableQueue(Long id) {
-		storeQueueAdminHelper.enableQueue(id);
-	}
-
-	public void disableQueue(Long id) {
-		storeQueueAdminHelper.disableQueue(id);
-	}
-
-	@Transactional
-	public void deleteStore(Long id) {
-		storeAdminHelper.deleteStore(id);
-	}
 }
