+++ store-service/src/main/java/com/boeingmerryho/business/storeservice/domain/repository/StoreRepository.java
@@ -1,18 +1,9 @@
 package com.boeingmerryho.business.storeservice.domain.repository;
 
-import java.util.List;
-import java.util.Optional;
-
 import com.boeingmerryho.business.storeservice.domain.entity.Store;
 
 public interface StoreRepository {
 	Store save(Store store);
 
 	boolean existsByStadiumIdAndNameAndIsDeletedFalse(Long stadiumId, String name);
-
-	Optional<Store> findById(Long id);
-
-	Optional<Store> findByIdAndIsDeletedFalse(Long id);
-
-	List<Store> findAllByIsDeletedFalse();
 }
