+++ store-service/src/main/java/com/boeingmerryho/business/storeservice/domain/service/StoreValidator.java
@@ -1,8 +1,7 @@
-package com.boeingmerryho.business.storeservice.infrastructure.helper;
+package com.boeingmerryho.business.storeservice.domain.service;
 
 import org.springframework.stereotype.Component;
 
-import com.boeingmerryho.business.storeservice.application.dto.request.StoreUpdateRequestServiceDto;
 import com.boeingmerryho.business.storeservice.domain.reader.StadiumReader;
 import com.boeingmerryho.business.storeservice.domain.repository.StoreRepository;
 import com.boeingmerryho.business.storeservice.exception.StoreErrorCode;
@@ -24,12 +23,6 @@ public class StoreValidator {
 
 		if (storeRepository.existsByStadiumIdAndNameAndIsDeletedFalse(stadiumId, name)) {
 			throw new GlobalException(StoreErrorCode.ALREADY_REGISTERED);
-		}
-	}
-
-	public void validateHasUpdatableFields(StoreUpdateRequestServiceDto dto) {
-		if (dto.name() == null && dto.openAt() == null && dto.closedAt() == null) {
-			throw new GlobalException(StoreErrorCode.NO_UPDATE_FIELDS_PROVIDED);
 		}
 	}
 }
