--- store-service/src/main/java/com/boeingmerryho/business/storeservice/infrastructure/config/QueryDslConfig.java
@@ -1,21 +1,0 @@
-package com.boeingmerryho.business.storeservice.infrastructure.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import com.querydsl.jpa.impl.JPAQueryFactory;
-
-import jakarta.persistence.EntityManager;
-import lombok.RequiredArgsConstructor;
-
-@Configuration
-@RequiredArgsConstructor
-public class QueryDslConfig {
-
-	private final EntityManager entityManager;
-
-	@Bean
-	public JPAQueryFactory jpaQueryFactory() {
-		return new JPAQueryFactory(entityManager);
-	}
-}
