--- store-service/src/main/java/com/boeingmerryho/business/storeservice/presentation/dto/request/StoreUpdateRequestDto.java
@@ -1,10 +1,0 @@
-package com.boeingmerryho.business.storeservice.presentation.dto.request;
-
-import java.time.LocalDateTime;
-
-public record StoreUpdateRequestDto(
-	String name,
-	LocalDateTime openAt,
-	LocalDateTime closedAt
-) {
-}
