--- store-service/src/main/java/com/boeingmerryho/business/storeservice/presentation/dto/response/StoreDetailResponseDto.java
@@ -1,14 +1,0 @@
-package com.boeingmerryho.business.storeservice.presentation.dto.response;
-
-import java.time.LocalDateTime;
-
-public record StoreDetailResponseDto(
-	Long id,
-	Long stadiumId,
-	String name,
-	LocalDateTime openAt,
-	LocalDateTime closedAt,
-	Boolean isClosed,
-	Boolean isQueueAvailable
-) {
-}
