--- store-service/src/main/java/com/boeingmerryho/business/storeservice/presentation/dto/response/StoreSearchAdminResponseDto.java
@@ -1,14 +1,0 @@
-package com.boeingmerryho.business.storeservice.presentation.dto.response;
-
-import java.time.LocalDateTime;
-
-public record StoreSearchAdminResponseDto(
-	Long id,
-	Long stadiumId,
-	String name,
-	LocalDateTime openAt,
-	LocalDateTime closedAt,
-	Boolean isClosed,
-	Boolean isDeleted
-) {
-}
