--- store-service/src/main/java/com/boeingmerryho/business/storeservice/presentation/dto/response/StoreSearchResponseDto.java
@@ -1,13 +1,0 @@
-package com.boeingmerryho.business.storeservice.presentation.dto.response;
-
-import java.time.LocalDateTime;
-
-public record StoreSearchResponseDto(
-	Long id,
-	Long stadiumId,
-	String name,
-	LocalDateTime openAt,
-	LocalDateTime closedAt,
-	Boolean isClosed
-) {
-}
