--- store-service/src/main/java/com/boeingmerryho/business/storeservice/presentation/dto/response/StoreUpdateResponseDto.java
@@ -1,13 +1,0 @@
-package com.boeingmerryho.business.storeservice.presentation.dto.response;
-
-import java.time.LocalDateTime;
-
-public record StoreUpdateResponseDto(
-	Long id,
-	Long stadiumId,
-	String name,
-	LocalDateTime openAt,
-	LocalDateTime closedAt,
-	Boolean isClosed
-) {
-}
