--- store-service/src/main/java/com/boeingmerryho/business/storeservice/utils/SortConstants.java
@@ -1,6 +1,0 @@
-package com.boeingmerryho.business.storeservice.utils;
-
-public class SortConstants {
-	public static final String CREATED_AT = "createdAt";
-	public static final String UPDATED_AT = "updatedAt";
-}
